+++ samples/separate/SubscriptionObject.cpp
@@ -3,7 +3,7 @@
 
 #include "TodayObjects.h"
 
-#include "graphqlservice/Introspection.h"
+#include "graphqlservice/introspection/Introspection.h"
 
 #include <algorithm>
 #include <functional>
