+++ samples/unified_nointrospection/TodaySchema.cpp
@@ -3,7 +3,7 @@
 
 #include "TodaySchema.h"
 
-#include "graphqlservice/Introspection.h"
+#include "graphqlservice/introspection/Introspection.h"
 
 #include <algorithm>
 #include <array>
